+++ extensions/wave/src/lib/gql/queries.ts
@@ -1,10 +1,10 @@
 import { NODES } from "./nodes";
 import { OBJECTS } from "./objects";
 
-export const queryGetCustomerOutstandingInvoicesCustomerStatement = `
+export const queryGetCustomerInvoicesCustomerStatement = `
     query($businessId: ID!, $customerId: ID!) {
         business(id: $businessId) {
-            invoices(customerId: $customerId, status: OVERDUE, sort: [INVOICE_DATE_ASC]) {
+            invoices(customerId: $customerId, sort: [INVOICE_DATE_ASC]) {
                 edges {
                     node {
                         ${NODES.Invoice}