+++ extensions/wave/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Wave Changelog
 
+## [] - {PR_MERGE_DATE}
+
+- Add Invoice Status to `tooltip` of Invoice `Icon`
+- Filter Invoices by Status
+- Improve "Outstanding Invoices" Customer Statement - now includes all invoices except: **PAID**, **DRAFT**, **OVERPAID**
+
 ## [Initial Version] - 2024-11-14
 
 - View Wave Businesses